requirements.txt
matplotlib==3.8.4
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
sage-importance==0.0.6
scikit-learn==1.5.0
seaborn==0.13.2
torch==2.4.0
torchmetrics==1.4.0.post0
torchvision==0.19.0
pytorch_lightning==2.3.0
tensorboard==2.18.0
tensorboard_data_server==0.7.0
zuko==1.3.0
PyYAML==6.0.1